Move classes to package org.robolectric.shadows.support.v4.
