+++ robolectric-shadows/shadows-support-v4/src/main/java/org/robolectric/shadows/support/v4/ShadowCursorLoader.java
@@ -1,4 +1,4 @@
-package org.robolectric.shadows;
+package org.robolectric.shadows.support.v4;
 
 import android.database.Cursor;
 import android.support.v4.content.CursorLoader;
