+++ robolectric-shadows/shadows-support-v4/src/main/java/org/robolectric/shadows/support/v4/ShadowDrawerLayout.java
@@ -1,9 +1,10 @@
-package org.robolectric.shadows;
+package org.robolectric.shadows.support.v4;
 
 import android.support.v4.widget.DrawerLayout;
 import org.robolectric.annotation.Implements;
 import org.robolectric.annotation.RealObject;
 import org.robolectric.annotation.Implementation;
+import org.robolectric.shadows.ShadowViewGroup;
 
 import static org.robolectric.internal.Shadow.directlyOn;
