+++ robolectric-shadows/shadows-support-v4/src/test/java/org/robolectric/shadows/support/v4/ShadowDrawerLayoutTest.java
@@ -1,4 +1,4 @@
-package org.robolectric.shadows;
+package org.robolectric.shadows.support.v4;
 
 import android.app.Activity;
 import android.support.v4.widget.DrawerLayout;
@@ -6,6 +6,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
 import org.robolectric.internal.ShadowExtractor;
+import org.robolectric.shadows.support.v4.ShadowDrawerLayout;
 import org.robolectric.util.TestRunnerWithManifest;
 
 import static org.assertj.core.api.Assertions.assertThat;
