+++ robolectric-shadows/shadows-support-v4/src/test/java/org/robolectric/shadows/support/v4/ShadowPagerAdapterTest.java
@@ -1,4 +1,4 @@
-package org.robolectric.shadows;
+package org.robolectric.shadows.support.v4;
 
 import android.database.DataSetObserver;
 import android.support.v4.view.PagerAdapter;
