+++ robolectric-shadows/shadows-support-v4/src/test/java/org/robolectric/shadows/support/v4/ShadowViewPagerTest.java
@@ -1,4 +1,4 @@
-package org.robolectric.shadows;
+package org.robolectric.shadows.support.v4;
 
 import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.ViewPager;
